rename /httpexception to /exception in preparation for merge
