+++ src/main/java/edu/byu/cs/controller/exception/InternalServerException.java
@@ -1,4 +1,4 @@
-package edu.byu.cs.controller.httpexception;
+package edu.byu.cs.controller.exception;
 
 public class InternalServerException extends Exception {
     public InternalServerException(String message, Throwable cause) {
