+++ src/main/java/edu/byu/cs/controller/exception/WordOfWisdomViolationException.java
@@ -1,4 +1,4 @@
-package edu.byu.cs.controller.httpexception;
+package edu.byu.cs.controller.exception;
 
 public class WordOfWisdomViolationException extends Exception {
     private static final String shortAndStout = "yo im a teapot";
