+++ src/main/java/edu/byu/cs/service/CasService.java
@@ -3,8 +3,8 @@ package edu.byu.cs.service;
 import com.fasterxml.jackson.dataformat.xml.XmlMapper;
 import edu.byu.cs.canvas.CanvasException;
 import edu.byu.cs.canvas.CanvasService;
-import edu.byu.cs.controller.httpexception.BadRequestException;
-import edu.byu.cs.controller.httpexception.InternalServerException;
+import edu.byu.cs.controller.exception.BadRequestException;
+import edu.byu.cs.controller.exception.InternalServerException;
 import edu.byu.cs.dataAccess.DaoService;
 import edu.byu.cs.dataAccess.DataAccessException;
 import edu.byu.cs.dataAccess.UserDao;
